r/consensus: update follower reply timestamp when received vote reply
When a raft group member is responsive i.e. it replied to the vote
request we should not skip sending it an append entries request.
Previously we did not updated the last received reply timestamp on vote
replies and thus it may happened that the newly elected leader would
have to skip sending append entries requests to the followers when
replicating configuration. This increased leader election latency as it
required sending the configuration batch to followers via
`recovery_stm`.

Fixes: #6675

Signed-off-by: Michal Maslanka <michal@redpanda.com>