storage: propagate is_internal_topic flag
This is a quick fix that propagates an ugly is_internal_topic
flag across many functions and classes to get it where it's
needed in segment_index and in copy_data_segment_reducer.

The neater solution will be to replace the raw filesystem paths
in classes like segment_reader with structured paths that carry
knowledge of the ntp & therefore carry knowledge of whether
the segment belongs to an internal topic.