s/ntp_config: make shadow indexing mode optional
This commit makes the shadow_indexing_mode field of storage::ntp_config
optional. This is required in order to be able to distinguish between
explicitly disabled shadow indexing at the topic level and the lack of
topic level configuration.

Previously, both topic configurations below resulted in the property
being set to model::shadow_indexing_mode::disabled.
* redpanda.remote.write=-1 && redpanda.remote.read=-1
* redpanda.remote.write=false && redpanda.remote.read=false

This was problematic
because the storage layer couldn't tell if the user explicitly disabled
shadow indexing via topic configuration or if the default shadow
indexing mode should be applied.