redpanda: plumb cluster discovery into application
This plumbs the cluster discovery module into `redpanda::application` as
one of the first subsystems initialized during startup.

Cluster discovery will have an initial dependency on the storage
subsystem, to get node UUID, cluster UUID, etc., and on the RPC
subsystem, to serve the RPCs required for the initial bootstrapping
sequence.

The rest of the subsystems will have a dependency on cluster discovery,
which, once automatic node ID assignment is supported, will be what
determines the node ID.

To that end, this commit orders the startup of the storage and RPC
subsystems up front, followed by cluster discovery, followed by the rest
of the subsystems.

This commit only plumbs cluster discovery infrastructure. The cluster
discovery sequence as a whole is unchanged.