archival: Fix off by one error in get_file_range
The check before this fix compared the file offset to file size using
greater-then operator. This is incorrect since it returns true in case
if file offset is equal to file size. This resulted in a situation when
we tried to read after EOF. In case of seastar::input_stream this
operation is allowed and returns 0 bytes. So the scan operation was
considered successful with the offsets from index.

(cherry picked from commit 858d2ef167195c86ea7b62810a7a93af83526001)