c/types: update SI mode on upgrade
Previously, the cluster level configuration for remote read and write
acted as overrides for the topic configuration. This means, for
instance,that even if remote write was disabled at the topic level it
could still be enabled via the cluster config. This has now changed in
the previous commits: cluster level configs are simply defaults at topic
creation time.

However, we need to handle upgrades for pre-existing topics. This commit
handles these scenarios by updating the deserialisation logic of
'topic_configuration'. The current cluster level config gets merged with
the topic level configs and that's treated as the source of truth from
now on.