tests: skip TimeQueryTest in debug mode
This test relies on reliably producing a burst of messages.

In debug mode docker tests, redpanda doesn't reliably hold leadership,
and the client can end up double-transmitting, violating the
test's expectation of receiving a set number of messages (nothing
to do with time querying, this is just the test's setup).

It's not yet clear if there is an underlying bug in idempotency,
or if this is a legitimate behavior.  In any case, stop running
this test in debug mode where it's known to be unstable.

Related: https://github.com/redpanda-data/redpanda/issues/6685
(cherry picked from commit 062b88785f5b3153cd639da0a3f3fd19b52f420d)