c/types: add total retention to topic properties
This commit introduces two new topic properties to
cluster::topic_properties: total_retention_bytes and total_retention_ms.
They should be used when remote write is enabled for the topic in order
to enfore total retention (i.e. retention policy that includes data
stored in the cloud and on disk).