c/types: disable total retention on upgrade
This commit changes the ADL and Serde deserialisation logic of
cluster::topic_configuration in order to handle upgrades.

Consider the following scenario: we are upgrading from a redpanda
version which is not aware of the retention.local-target.* topic
configuration options. If the topic in question is enabled for remote
writes to cloud storage, the existing retention.* topic configs control
the local retention. In the new version, for topics with remote write
enabled, retention.local-target.* should control local retention instead
and retention.* should be used for total retention.

To preserve behaviour for existing topics we disable the total_retention
topic properties when reading an old message.