kafka: enable altering local-target topic configs
This commit extends the handling of the IncrementalAlterConfigs Kafka
request to allow for altering the values of local-target topic
configuration properties.

Note that the ADL encoding logic for cluster::incremental_topic_updates
has been updated to include the new retention.local-target fields and
batch_max_bytes. This is required in the following scenario:
1. The IncrementalAlterConfigs Kafka request lands on a node that's not
   the controller leader. In this case, the broker sends an
   update_topic_properties RPC to the controller leader.
2. The update_topic_properites RPC is the first request sent on the
   connection, which means that it will be encoded in ADL regardless of
   the fact that both the client and the server support Serde.

Prior to this commit, in the described scenario, an update of
batch_max_bytes would not have been performed and the broker would have
not returned an error to the client.