c/feature_manager: init node versions from cluster_discovery
Peer seed nodes discovered throught cluster_discovery before the cluster
is bootstrapped are required to be at the same latest logical versions
as the local. Therefore as the cluster is initialized, versions of seed
nodes can safely be assumed to be at the latest. That enables the auto
node_id feature that is essential for cluster bootstrap.