test_utils: adds PUT support for http imposter
when a PUT request is sent, the content is stored for the URL and
returned in response to future GET requests for the same URL. This is
the same behavior as s3 imposter fixture.

Also adds support to fail the nth request to support tests where we want
to assert behavior when a number of similar requests are issued but one
of them fails.