cluster: Write fence batch on each begin_tx
tx_fence batch type is used to fence off a producer with older epoch.
We're repurposing it to act as a txn begin marker to store tx_seq (a
transaction id) and the txn timeout to get rid of the prepare phase.

It's dangerous to change data format in a distributed system because
on an upgrade we'll get a heterogeneous cluster and old node won't
understand the new format. To avoid this problem we switch to using
new version only when all nodes are upgraded (it's controller by the
transaction ga feature flag)

With getting rid of the prepare phase we still need to store
tx_seq and timeout, repurposing tx fence to use it as a txn
begin marker and to store those info.