cluster: Add tx_seqs map to log_state
We used to write txn records with acks=1 and Redpanda wasn't able to
tolerate leadership changes so it was fine to store auxiliary data
in memory. With changing acks from 1 to all we stepping from storing
data in memory and use log backed structure (log_state) for storing
things like txseqs.