net: avoid references to `this` capture in lambda coro
It still is not clear that capturing `this` in a lambda coroutine is
safe. The accept handler in the net server captures `this`.
Additionally, we've been dealing for some time with this crash:

```
TRACE 2022-10-13 08:55:06,465 [shard 0] rpc - conn_quota.cc:118 - put(172.16.16.27)
TRACE 2022-10-13 08:55:06,465 [shard 0] rpc - conn_quota.cc:130 - leaving put(172.16.16.27)
AddressSanitizer:DEADLYSIGNAL
=================================================================
==16388==ERROR: AddressSanitizer: SEGV on unknown address (pc 0x7fa76bfb218f bp 0x7ffddbd1a140 sp 0x7ffddbd19ea0 T0)
==16388==The signal is caused by a READ memory access.
==16388==Hint: this fault was caused by a dereference of a high value address (see register values below).  Disassemble the provided pc to learn which register was used.
AddressSanitizer:DEADLYSIGNAL
AddressSanitizer: nested bug in the same thread, aborting.
```

The only place where:

```
TRACE 2022-10-13 08:55:06,465 [shard 0] rpc - conn_quota.cc:130 - leaving put(172.16.16.27)
```

This is happening is in and around this lambda. However, nothing else seems
obviously problematic which might lead to dereferencing a junk pointer. So this change
is not definitely going to fix it as we don't really have proof that the problem is `this`
capture being clobbered.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>