tests: fix SISettings handling in AWS
The SISettings object was only getting adjusted for
using AWS S3 (via `load_context`) when provided at
construction time, not via set_si_settings.

This caused some tests not to work properly
when run on EC2 instances.

Move the call to a location where it'll get hit
in all cases.