cloud_storage: Update partition_manifest
Add 'replaced' collection of segments to the manifest. The collection
contains segments which were re-uploaded.

Also, extend segment_meta struct by adding version, delta_offset_end,
and segment_term fields. The 'version' field contains segment_meta
version. If the field is not in JSON it's assumed to be v1 and
everything works as before. Otherwise the version is set to v2 (the only
other opiton). New methods that generate segment name/path were added
which are taking the segment_meta as a parameter. In case of v2 meta
they will generate slightly different names. The name will contain not
only base offset but also a committed offset. This is needed to avoid
overwriting existing segments when the compacted segments are
re-uploaded.

The 'delta_offset_end' field contains a delta offset for the last offset
in the segment. This could be used to simplify some offset translation
logic. The 'segment_term' is the term during which the segment was
generated (the same as in the segment name).

The generate_segment_name function was renamed to
generate_local_segment_name because its result doesn't always match the
segment name in the cloud.