test: look for llvm-symbolizer when running tests
Before we were relying on seastar to decode backtrace for asan and
friends. For whatever reason that decoding isn't currently working. It
looks like the only reason that lsan wasn't doing this for us already
was that no symbolizer was available. Since we have one already with our
llvm build, we can just use that.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>