redpanda/application: add checks needed for storage_strict_data_init
When the `storage_strict_data_init` property is enabled a user will
have to manually add an empty magic file called `.redpanda_data_dir`
to Redpanda's data directory for RP to start. RP currently has no way
of telling if it's data directory is purposely or accidently empty.
Hence if an incorrect volume or filesystem is mounted RP will initialize
a data dir in it incorrectly. With `store_strict_data_init` enabled RP
will instead look for the magic file on start and exit if it's not there.