cluster: Change add_segments method in ...
the archival_metadata_stm. Instead of passing the partition_manifest
and compute the difference pass the manifest that contains the new
segments. This is needed to simplify partition manifest and to remove
difference method.

The partition_manifest is used isntead of vector of segments to simplify
code maintenence. The partition_manifest has a forward declaration so we
don't need to include the header in archival_metadata_stm.h.