cloud_storage: Extract segment_meta to types.h
The segment_meta is an internal type of the partition_manifest but it's
used outside the manifest pretty often. This commit moves the struct to
'cloud_strage/types.h'. It also changes the 'replaced_segments' method
to return 'segment_meta' instead of 'segment_name_meta' (which is
removed). This is possible since we guaranteed to always have a
'segment_meta::segment_term' field set to proper value by the manifest.
Because of that we're no longer need to return a segment name.