v/cloud_storage/cache_service.cc: fix a potential race condition whereby a file becomes a candidate for deletion and stat is requested for it.
Under normal operations, deletion would occur sometime after stat is
requested.
But if a lot of these files are generated closely in time, the deletion
could happen for a file before stat is requested for it.

This change request stat before making the file a candidate for
deletion, removing this potential race.