r/consensus: do not require learner promotion to leave joint consensus
In redpanda Raft implementation reconfiguration cancellation is done by
reversing the direction of configuration change. When Raft group
configuration change is in first phase i.e. new nodes are added as
learners to current configuration then they are simply removed.
Cancellation of change when reconfiguration entered a Joint state
requires swapping old and new configurations in Joint raft group
configuration. It may be the case that cancellation will never finish
even if only one node is unavailable as the node may be a voter that was
demoted to learner in the last step before its removal.

In order to allow the configuration change to finish we allow Raft to
leave joint consensus before all learners are promoted to voters. This
change is safe as learners does not change the safety guarantees but
enables us to reliably cancel partition movement when one of the nodes
is down.

Signed-off-by: Michal Maslanka <michal@redpanda.com>