cluster/storage: put remote_delete property into ntp_config
This simplifies the implementation of cluster::partition, as it
can use the existing ntp_config object to carry the knowledge of
whether a topic's S3 data should be deleted on topic deletion.  The
alternative would be a proliferation of boolean function arguments
to carry the property around.

Our long term plan for ntp_config is that its should diverge
into subsets for local storage and remote storage, but currently
it contains most cloud storage properties.