rpk: make bootstrap set cluster formation config
For backwards compatibility, Redpanda's default is to continue to use
the old style of startup. New clusters should should use the new style
of bootstrapping, where all nodes are provided the same set of seed
servers. To that end, this updates the bootstrap command to set the
`empty_seed_starts_cluster` node config to false.