rpc/transport: move `as_scattered` call from `dispatch_send` to `do_send`
Adding additional scheduling points in `dispatch_send` to account for
the newly futurized `as_scattered` results in a race condition where
messages are sent out of order. As such the easiest solution is to call
`as_scattered` outside of `dispatch_send` where adding a scheduling point
won't result in this race condition. Then pass the result to `dispatch_send`.