tests/pp: add stress test for client stop
The kafka client stop() method is called whenever a client is evicted
from the LRU cache. This commit adds a ducktape test that configures the
cache to a max size of 1. The test stresses client stop() by issuing
many rest requests with two unique users within a 30s loop. The idea is
tigger eviction, and consequently client stop(), many times.

Closes #6595

This commit also adds options to the RedpandaService for the client
cache keep alive and client cache max size configs.