txns/compaction: short circuit tx_reducer for non tx topics
For non transaction topics (no rm_stm attached), it does not make
sense to run tx_reducer logic as we won't find any batches of interest.

While these topics typically build their compaction index along with
the data segment writes, in certain rare cases the index can go missing/
corrupt and requires a rebuild. In such cases we can short circuit
tx_reducer and delegate it to the underlying reducer.

We found this issue when __consumer_offsets topic was trying to rebuild
it's index (probably after a crash) and the reducer came across some
unexpected batches that triggered an assert.