rpk: modify kafka client initialization, revert fd7a66c385d88882bd819c779eb2734173011eeb
The recent bump in franz-go returns a better error if the first request
fails with no sasl. This is related to #1944, but now that franz-go
returns a fully worded error directly, we no longer need to "ping"
redpanda with a metadata request when we initialize a client (i.e. we
revert #4727). We also were only pinging for admin clients, and not
everywhere uses an admin client (yet), so the prior implementation
wasn't entirely comprehensive (we all missed this).

Importantly, removing this metadata request makes admin initialization
behave similarly to non-admin client initialization (#6317) The previous
difference is why `rpk topic list` worked, and `rpk topic alter-config`
did not: the former used the admin client, the latter does not.

Now that clients are initialized the same, we also tighten some
timeouts. See the embedded comments for why. This might allow #6367 to
be reverted entirely.

In this same commit, we revert fd7a66c385d88882bd819c779eb2734173011eeb.
"""
Revert "tests/admin_op_fuzzer: add custom shuffling of the rpk brokers list"
"""
fd7a66c was introduced to work around the bug that this commit is
fixing. We will revert the hack-around-the-bug changes in test code to
indirectly test this commit.

Closes #6317.

(cherry picked from commit d0637423e648af7c7461167613f1f8c6272abdd9)