rpk: modify kafka client initialization
The recent bump in franz-go returns a better error if the first request
fails with no sasl. This is related to #1944, but now that franz-go
returns a fully worded error directly, we no longer need to "ping"
redpanda with a metadata request when we initialize a client (i.e. we
revert #4727). We also were only pinging for admin clients, and not
everywhere uses an admin client (yet), so the prior implementation
wasn't entirely comprehensive (we all missed this).

Importantly, removing this metadata request makes admin initialization
behave similarly to non-admin client initialization (#6317) The previous
difference is why `rpk topic list` worked, and `rpk topic alter-config`
did not: the former used the admin client, the latter does not.

Now that clients are initialized the same, we also tighten some
timeouts. See the embedded comments for why. This might allow #6367 to
be reverted entirely.

In the commit we are cherry-picking from, we revert fd7a66c, but that
does not exist in v22.1.x so we have nothing more to do here.

Closes #6317.

(cherry picked from commit d0637423e648af7c7461167613f1f8c6272abdd9)
(cherry picked from commit 906a76889e79737a3f539ce154b9d14444674d86)