rpk cluster config get: avoid scientific notation
The scientific notation was from json unmarshal rules. We now convert to
int64 directly.

Note that edit, import, and export all handle this already.

We remove the one impossible branch in export, since we know we cannot
have an `int` in the config map, and we remove the hack in the test that
converted from scientific notation.

Closes #6070.