config: enables config edit when using IAM roles
When IAM roles are used, cloud storage keys are expected to not be
supplied. This set of changes fixes an issue where keys were expected to
be present when IAM roles were used.

Now when IAM roles are active keys should not be supplied, and when
static auth is used keys must be present.