cloud_storage: log errors during remote startup
when remote starts with IAM roles it launches a background op which
refreshes credentials. if there are unhandled exceptions during this
operation they cause a crash in redpanda. this change makes sure that
the errors are caught and logged and the system does not crash.

(cherry picked from commit 67375b0ed54e4a88cd54d3701997eba9467e7333)