cluster: Change signature of the add_segments ...
method of the archival_metadata_stm. Pass vector of segment_meta structs
instead of the partition manifest.

Previously, the method accepted parition_manifest which was used as a
collection of segments. This is error prone because it's easy to pass
the whole manifest. The state of the STM is not idempotent so all
segments will be added one more time. This change fixes this API. Also,
note that the 'segment_meta' is now required to have 'segment_term' to
be set to some meaningful value.