rm_stm: support txn coordinator forgetfulness
Txn coordinator saves state diffs to log on every commit but
keeps the ongoing transactions in memory so in case of a reboot
it rolls back in time to the moment of the last committed tx.

From the rm_stm perspective it looks like the tx coordinator
forgets about the ongoing txes and may replay the latest com-
mitted tx for any pid.

Updating commit_tx to support this scenario.