txns/tm_stm: Add topic revision to partition state
This commit adds the topic revision_id to the transaction
state that is maintained on the coordinator. This revision is returned
by the partition itself as a part of the begin_tx RPC.

A topic's revision is the offset of the create command for that topic
in the controller log. It is monotonically increasing. We rely on these
properties in the next commit to identify if a topic and it's partition
are deleted [and recreated] in the middle of a transaction.