c/topic_table: topic lookup by revision id
Use case: For transactions we need a reliable way of determining if a
topic exists. We are okay with false positives (as the transactional
RPC will anyway fail if the ntp does not exist) but false negatives are
dangerous because we skip committing to the topics that we think are
missing. A particularly tricky case with false negatives is when the
controller replay is still in progress after a recovery. If a
trasnactional RPC comes in while the replay is in progress, just looking
up in the cache can yeild false negatives.

This commit adds utilities that avoid these false negatives and builds
on top of the topic revision added to coordinator state in the previous
commit.

The contains() check added in this commit can return 3 possible outcomes

(1) indeterminate - the cache is not sufficiently up to date.
(2) exists - the cache has topic + revision
(3) not_exists - all other cases, we are sure the topic revision is deleted.

It is to be noted that (2) can yeild false positivies if the replay is
still in progress but we are okay with that.