tests: retrofit to use Verifiable{Producer,Consumer}
I intend on backporting UpgradeBackToBackTest to v22.2.x.
UpgradeBackToBackTest uses kgo-verifier, which has been heavily updated
on dev since v22.2. Rather than backporting the infra changes and
corresponding updates to usage in tests, this commit swaps out the
test's usage for the corresponding Verifiable* variants.

Also adjusted workloads to not use idempotence.

This commit is not meant to be merged to dev!