tests: enable debug logging for pandaproxy
Recently the test tree has had an influx of ducktape tests for the
Pandaproxy and Schema Registry. This is in part due to increased
customer demand for these services. We are starting to see more CI
failures related to the proxy and registry because of the increase in
test coverage. Therefore, let's enable debug logging by default.