redpanda: early load of feature table snapshot
This enables all later subsystems to operate on the assumption
that outside of mixed version systems, they can count on
having their feature flags set from the moment they start.

This enables things like the RPC layer to use new-version
protocols robustly, and avoid the complexity of sometimes
having to run in a legacy mode if the controller didn't
finish replaying the feature table updates yet.