rpk config: better defaults for empty rpk brokers; favor non-TLS
The prior commits introduce behavior to use redpanda.admin if
rpk.admin_api is empty. We previously had this same behavior for
rpk.kafka_api, but this only used redpanda.kafka_api[0]. We abstract the
logic from the previous commit (and have a authn => no authn helper) to
make using the prior commit just as simple for brokers.

We also do *not* use TLS listeners if we have any non-TLS listeners: rpk
likely cannot fill client certs, and they may not be on the host. The
expected behavior is that a person ssh's into the node and then runs
rpk; localhost non-TLS communication will work.

The new abstraction is well setup for when we add support for pandaproxy
and schemaregistry to rpk, both of which use the same types.

This adds testing for all this behavior as well.