cluster: add join_type to join_node_request
This commit adds an enum to the join_node_request to differentiate when
callers want to register with (simply record their node UUID, potentially
being assigned a node ID) or join (be added to the controller Raft
group, begin receiving partitions) a cluster.

Previously this was inferred based on whether or not a node ID is
provided, indicating a join. When no node ID was provided, the
members_manager would only register the node UUID and return, without
adding the node to the controller group.

I intend on having a registration of a node UUID that doesn't entail
joining: upon upgrading a cluster, the nodes' UUIDs need to be
registered, even though the nodes have already joined the cluster.