rpk debug bundle: always choose first host
admin api's PrometheusMetrics actually requires one host to send to
(i.e. you must pick the one host you want metrics from). Our tests
previously passed because rpk.admin_api was empty and then we used the
default 127.0.0.1:9644. The tests now failed because redpanda.admin was
non-empty and we picked all three brokers in it. Because we favor the
most-likely-accessible host first, we can just pick the first host in
the config.