bytes: move oncore checks out of header
The oncore checks require access to an expensive seastar header
(core/smp.hh), as well as vassert.h which is expensive due to fmt.
However, oncore checks are only enabled on debug builds. Therefore we
can move these checks out of the header as they offer no real-world
performance penalty when incurring the extra function call.

Note that keeping oncore in v::bytes is a bit silly (it is used
elsewhere). But it is the most convenient place to stash it given the
current set of dependencies (e.g. it can't live in utils). Ultimately I
believe we should merge bytes and utils libraries.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>