tests: avoids selecting same node for disruptions
The action injector selects the next node for action using randomness
from a set of available nodes. For reversible actions, the node which
was affected is added back to available nodes once the action is
reversed.

This leads to a situation where the same node can be selected again and
again for disruptions, especially when the node count is low and so
randomness is not as effective.

This change makes sure that once a node is selected for an action, it is
not selected again until all of its peers have been processed. A set of
nodes affected is maintained and these are removed from nodes available
for action.

Once all nodes are processed, the set of affected nodes (which is now
full) and the set of available nodes (which is now empty) are swapped,
beginning a new cycle of actions.