cluster: Add use int32_t for tx_status enum
tx_status for transaction struct was not specified, so it may create some problems.
Because we serialize it with adl. And if basic int type (for enum) was changed,
redpanda can not deserialize old log records.
To avoid this we need to put type which compiler uses for this enum.

std::is_unsigned<decltype(std::__to_underlying<tm_transaction::tx_status>)>::value
shows false
sizeof(tm_transaction::tx_status) is 4
so it is int32_t