redpanda_installer: clean up after download failure
This commit adds some basic cleanup when failing to download, e.g. if
the network was down while the test run. This previously borked the rest
of the upgrade tests in a given test run, as it left remnants of the
botched installation.

There are more improvements that can be done in this area (breaking up
the single bash command that installs into several, handling failures on
a per-node basis, etc), but they typically entail untangling some
parallelism for dedicated noes. This commit implements a basic cleanup.

I tested this manually by running an upgrade test while disabling my
Internet connection temporarily. Upon completion of the test, the
`redpanda_installs` directory didn't have the botched files and I could
run the test again successfully upon reenabling the connection.