pp/cache: mv sharded state into sharded service
Prior to this commit, all sharded state was handled in the
sharded_client_cache wrapper. That wrapper existed on a single core
which led to concurrency failures since there was cross-core
communication betweem the wrapper and the sharded services.

This commit reduces cross-core communication by moving concurrency
mechanisms into the sharded service. This commit also serves as the
pre-req to remove the sharded_client_cache wrapper.

Finally, the auth_ctx_server is the new "frontend" for the kafka client
cache. The ctx server can pass function handles to the sharded instance
directly. So invoke_on_cache is no longer needed.