tests: increase sleep interval on cycle complete
Once all nodes have had a disruptive action applied to them, a new node
is selected from the same set for the next disruptive action. There is a
chance that the same node which was last in the previous cycle will be
selected again, thus running two successive actions on the same node.
This increases the chance of a test failure. Disruptive actions are
applied to random nodes but we do not expect the same node to keep
getting restarted in succession. To avoid this once, the cycle is
complete, the sleep interval is temporarily increased, so that even if
the same node is selected it has a chance to recover.