cloud_storage: Support destructive manifest ...
operations in remote_partition. Truncation and/or replacement of data
in the cloud may happen asynchrously with the fetch operation. In this
case we have to rebuild the collection of segments inside the
remote_partition. This operations causes all materialized segments to be
offloaded and all cached readers to be evicted. The data stays cached
inside the cloud storage cache.