cloud_roles: ensures http client is always closed
The http client in cloud_roles::make_request was not closed when
exception is thrown. RAII might usually be a better fit for ensuring the
client is closed, but the client stop method returns a future which
cannot be returned from/blocked on in destructor.

This change introduces a helper modeled after ss::with_file to accept
and own an http client and close it after a user-supplied operation is
finished.

It also adds a helper to wrap making an http request with catching and
logging common http call errors.