kafka: simulate a unity build with code generation
A significant space / time overhead went into producing each individual
object archive for message response/request types (66 in total). This
was to the point that each one was 5-10 MB in size for debug builds.
There was a huge amount of duplication between these object files.

This commit adapts the generator to combine all sources into a single
file. The result is a 10x reduction in object file size.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>