rpk redpanda config bootstrap: improve initialization
This command previously always set certain fields in the redpanda.yaml.
We now change the command to only set values that are current default
values, or if values are empty. Manually changed values will be left
unchanged.

One alternative that we want but cannot proceed with is to bail if
redpanda.yaml is already present. We have many existing flows of config
parameters being set before bootstrap is called, so bailing on an
existing file would be a breaking change.

We modify the tests in this PR to be strictly file based: this is easier
to correct in the future, easier to add tests to, and is more
comprehensive.

Closes #6495.