cluster: fix archival_metadata_stm collectible offset logic
This is a subtle upgrade case.  When a 22.3 node starts and
has not yet executed the migration to move tiered storage
enablement flags from global defaults onto the topic, the
archival_metadata_stm::max_collectible_offset method
will naively return offset::max (i.e. giving storage-gc
permission to proceed without waiting for uploads).

Fix this by reinstating the legacy behavior (of assuming
tiered storage is enabled if the global default is true)
conditionally in the case where we are in the middle
of an upgrade (i.e. when 22.3 feature flag is not
yet active).