cluster_discovery: assume non-founder if controller exists
Currently we can exit out of cluster discovery if we discover that the
local node isn't a cluster founder. The list of reasons to exit early is
currently:
- we are not in the seed server list
- we have a cluster UUID in our key-value store
- another seed server already advertises a cluster UUID

These alone are sufficient in prevent accidental bootstrap in the event
of a node wipe. However, these alone are insufficient for performing a
rolling restart into using the seeds-driven configs (e.g. if an
automated deployment started using the new configs automatically after
performing an upgrade).

Consider a rolling restart where before the restart, all nodes were
configured with root-driven bootstrap (empty_seed_starts_cluster=true).
At a certain point, there would be a mismatch between the restarted node
and the other nodes, and Redpanda will throw an error.

To avoid this, this commit adds a check during cluster discovery to
assume that we are a non-founder if a non-empty controller directory
exists. With this, there is less confusion around mismatches in remote
configurations, since the only time remote state is relevant is when
there is no local state.