cloud_storage: Avoid offloading materialized segments
When the remote_partition synchonizes with the manifest all materialized
segments are offloaded. This creates several problems. The performance
may suffer if the updates to the manifest are frequent. After every
update we will offload existing materialized segments. This will require
all readers to be re-created. The segments could still be in cache but
the in-memory state will be lost.

The second problem is caused by the remote_segment limitation. The
remote_segment uses cache service under the hood. The instance of
remote_segment expects to be the only instance that has access to the
segment in the cache. If we will have two instances of the
remote_segment for the same log segment in memory they may conflict over
the data in the cache. The offloading process is lazy. When the
remote_segment is offloaded it's moved to the evicted list. The items
from the list are pulled out by the background fiber and stopped.

This makes the following situation possible: the reader starts fetching
data from the segment. The remote_segment is just created and started
hydrating segment on disk. Then the manifest gets updated and this
causes the sync method to move remote_segment to the eviction list.
Then the client materializes the segment and second instance of the
remote_segment is created. This instance is trying to hydrate a segment
by hydration is already in progres. This breaks invariant of the
remote_segment and triggers assertions.

This commit solves the problem by avoiding offloading of materialized
segments which are not changed in the manifest. The sync method first
pulls all materialized segments to the temporary map. Then it rebuilds
_segments collection but instead of creating offloaded_segment_state
instances for every segment it checks the map. If the map contains
materialized segment for that particular segment the materialized
segment is used.

The remaining materialized segments which are left unused are offloaded.
These segments are replaced by the compacted segments and can be safely
moved to the eviction list.