r/group_manager: prevent replicas from becoming a leader when not ready
When redpanda starts it may take a while before Kafka API listener is up
and running. During that time raft groups replicas that are already
instantiated may become leaders. Being a leader makes Kafka requests to
be sent to the leader node (f.e. topic operations to controller,
describe requests to partition leader, etc.). When node is not yet ready
to serve Kafka API requests but it already has some partition leaders it
may disrupt the clients as they will not be able to connect to a node.

When redpanda node is not ready to serve Kafka API requests we should
prevent raft groups from becoming a leader. This prevents an
interruption of clients as they will keep contacting other cluster nodes
which are available.

Fixes: #6892

Signed-off-by: Michal Maslanka <michal@redpanda.com>