cluster: Add clear methods for tx pids
We need to prevent overflow for some
state maps, so we need to delete info
for pids which does not have tx.
For this we just should check is pid
inside tx_seqs map or not. If not it means
tx was commited or aborted

(cherry picked from commit 2379bbd1f29a3277541382a06ea2b179a44b8fff)