transactions: fix check for ongoing tx
Abort is idempotent, basically if it can't find an ongoing tx Redpanda
thinks that it was already aborted and returns success without doing
anything. The problem is that when we check for an ongoing tx we check
if a partition already got any data but ignores if a transaction started
but hasn't received any data. Fixing it.