s3: ensures that stream is closed when uploading segment
In s3 client put_object, an input stream is moved into the function and
expected to be unconditionally closed by the function. There is a
chance that the function body throws an exception and the stream is
never closed.

With the changes to support uploading compacted segments, the stream is
actually a wrapper for a file data source, not closing this can trigger
an assertion and crash the application.

This change ensures that the relevant parts of the code are called with
seastar futurize_invoke to make sure that the stream is closed.