c/controller_backend: always update leaders ...
table on deletion

The partition leaders table contains partition leaders for all partitions
accross the cluster. For this reason we need to delete from it regardless
of whether a replica of 'ntp' is present on the node, if performing a
partition deletion (i.e. deletion mode is global).

Currently, the table is only updated if a replica of the partition being
deleted is present on the host. Thus, if the replication factor of a
partitions is not equal to the size of the cluster, the nodes that are
not replicating the partition will not delete it from their local
tables.

One reason why this is problematic is because the cluster level
metrics published from 'controller_probe' use the partition leaders
table.