tx/compaction: Adjust aborted tx range for compacted segments.
This commit adjusts the begin offset from which we consider aborted
transactions for a given fetch. For non-(self)-compacted segments, this
doesn't matter as the client knows how to align the aborted segment
offsets with the data batches but it matters for self-compacted
segments.

A typical client (modeled after Kafka) does the following to apply the
aborted ranges for data batches.

---- consume logic ------

for each data batch:

  // refresh running list of aborted txns
  consume_aborted_offsets_up_to(batch.end_offset)

  // Check if the current pid is aborted
  if is_transactional(batch) && pid(batch) in aborted:
    // ignore batch

  // Update running list of aborted txns
  if is_aborted_control_batch(batch):
     // remove pid from running aborted txn list

-------------------------

If the entire transaction falls with a self compacted segment, we
override the transactonal bit for every batch making it a non
transactional batch, so none of the aborted list is applied.

However if a transaction spans a self compacted and an uncompacted
segment. For example

--- before compaction ---
begin_tx -- segment 0
data_batch_1 -- segment 0
data_batch_2 -- segment 1
commit_tx -- segment 1
------------------------

and only segment 0 is self compacted, we are left with the following

--- after compaction ----
data_batch_1 (non-tx) - segment-0
data_batch_2 (tx) - segment-1
commit_tx -- segment-1
------------------------

This is particularly possible if segment-1 is the active segment that we
do not compact until closed.

So applying the same consumption client logic for the entire set of
batches from begin to end fails as we do not have abort markers in
segment 0 that are needed to keep the aborted txns list up to date.

So this patch essentially fixes the issue by populating aborted txn
offsets from the first transactional offset begin, which is data_batch_2
in segment-1 in the above example.

Some notes for future:

An _ideal_ way to fix this issue is by updating aborted transactions
list along with compaction. With our current design we decouple the
aborted transaction offsets from segments, so there is no atomic way
of swapping both together at the same time.